Split equivalency tests
